# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Filmes",
    "version": "1.0.0",
    "description": "API para gerenciamento de filmes utilizadas nas disciplinas de desenvolvimento Web no IFPI Campus Picos"
  },
  "paths": {
    "/api/movies": {
      "post": {
        "summary": "Adiciona um novo filme",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Filme criado com sucesso"
          }
        }
      },
      "get": {
        "summary": "Retorna uma lista de todos os filmes",
        "responses": {
          "200": {
            "description": "Lista de filmes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/movies/{id}": {
      "get": {
        "summary": "Retorna um filme específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filme retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza as informações de um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filme atualizado com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Remove um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Filme removido com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "O Poderoso Chefão"
          },
          "description": {
            "type": "string",
            "example": "Um épico da máfia focado na família Corleone."
          },
          "releaseYear": {
            "type": "integer",
            "example": 1972
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-07-19T12:34:56Z"
          }
        },
        "required": ["id", "title", "description", "releaseYear"]
      }
    }
  }
}